Update Carousel to imitate the one in netflix
I realized that the netflix's carousel didn't show the last element next to the first.
e.g. if our carousel had 10 item and we showed 3 per slide, we would have: first cards 1, 2, 3, then: 4, 5, 6, then: 7, 8, 9. The difference came when we hit the next button again. In netflix's one it moves only 1 card, showing 8, 9, 10, but it mine it moved 3 showing 10, 1, 2.